requirements.txt
streamlit==1.28.0
langchain==0.1.0
langchain-openai==0.0.5
langchain-google-genai==0.0.6
requests==2.31.0
google-auth==2.23.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.1
google-api-python-client==2.0.2
python-dotenv==1.0.0
pytz==2023.3
